2011-11-22  Christian Betz  <cbetz@sicompos.com>
    merge fix for long-long on windows (by Zoltán Böszörményi)
    requires misinttypes for Visual Studio (http://code.google.com/p/msinttypes/)

    * configure.in:
    * libsrc/formatstring.c (rlib_number_sprintf),
    (rlib_format_string_default):
    * libsrc/fxp.c (rlib_fxp_div):
    * libsrc/pcode.c (rlib_pcode_dump):
    * libsrc/pcode.h:
    * libsrc/pcode_op_functions.c (rlib_pcode_operator_str):
    * libsrc/util.c (rlib_safe_atoll):
    * libsrc/value.c (rlib_var_get_number):

+++ libsrc/pcode.h
@@ -42,7 +42,7 @@
 #define RLIB_FXP_MUL(a, b) rlib_fxp_mul(a, b, RLIB_DECIMAL_PRECISION)
 #define RLIB_FXP_DIV(num, denom) rlib_fxp_div(num, denom, RLIB_FXP_PRECISION)
 
-#define RLIB_FXP_TO_NORMAL_LONG_LONG(a) (a/RLIB_DECIMAL_PRECISION)
+#define RLIB_FXP_TO_NORMAL_LONG_LONG(a) ((gint64)(a/RLIB_DECIMAL_PRECISION))
 #define LONG_TO_FXP_NUMBER(a) ((a)*RLIB_DECIMAL_PRECISION)
 
 #define OP_ADD		1
